+++ libs/gesha-core/src/conversions/mod.rs
@@ -1,8 +1,8 @@
 mod error;
 pub use error::{Error, Result};
 
-mod definition;
-pub use definition::{Converter, TestDefinition};
+mod converter;
+pub use converter::{Converter, TestDefinition};
 
 mod test_case;
 pub use test_case::TestCase;
